# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API documentation for the library system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/getRentalReadingProgress/{rentalId}": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get reading progress for a rental",
                "operationId": "a3dbc70ea464fd84e9d2877c5d579838",
                "parameters": [
                    {
                        "name": "rentalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/rentBooks": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Rent a book",
                "operationId": "59a2f1e4cb5340d9387025af2e509b11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book rented"
                    },
                    "422": {
                        "description": "Book not available or validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/rentExtend/{rentalId}": {
            "patch": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Extend a rental",
                "operationId": "b9ad6d228149c1eee013989dde4df6fe",
                "parameters": [
                    {
                        "name": "rentalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "extendedDate"
                                ],
                                "properties": {
                                    "extendedDate": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Extended"
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "409": {
                        "description": "Rental already finished"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/updateRentProgress/{rentalId}": {
            "patch": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Update reading progress",
                "operationId": "c6d40ada9af88299570401b475f77822",
                "parameters": [
                    {
                        "name": "rentalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "progress"
                                ],
                                "properties": {
                                    "progress": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Progress updated"
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/rentFinish/{rentalId}": {
            "patch": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Finish a rental",
                "operationId": "b13797d1460b18309787c18b2aa33c7c",
                "parameters": [
                    {
                        "name": "rentalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental finished"
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "409": {
                        "description": "Rental already finished"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}